+++ src/test/resources/projects/define/test/define.js
@@ -1,0 +1,16 @@
+window.userAgent = {};
+
+/** @define {boolean} */
+const DEBUG = false;
+
+/** @define {string} */
+const STRING = "";
+
+/** @define {string} */
+const STRING2 = "";
+
+/** @define {number} */
+const NUMBER = 0;
+
+/** @define {boolean} */
+window.userAgent.ASSUME_IE = "false";